+++ Assets/CompanyClasses/InventoryEntry.cs
@@ -2,9 +2,9 @@ public class InventoryEntry
 {
     public Good good;
     public int quantity;
-    public float acquisition_price;
+    public decimal acquisition_price;
 
-    public InventoryEntry(Good good, int quantity, float acquisition_price)
+    public InventoryEntry(Good good, int quantity, decimal acquisition_price)
     {
         this.good = good;
         this.quantity = quantity;
