+++ Assets/CompanyClasses/iCompany.cs
@@ -1,9 +1,9 @@
 public interface iCompany
 {
     string company_name{get;set;}
-    void BuyGood(Good good, int quantity, float price,int period);
-    void SellGood(Good good, int quantity, float price,int period);
+    void BuyGood(Good good, int quantity, decimal price,int period);
+    void SellGood(Good good, int quantity, decimal price,int period);
     void Initialize(string company_name, CompanyLevelEnum company_level);
     Inventory GetInventory();
-    float Get_cash();
+    decimal Get_cash();
 }
