+++ Assets/Goods/iPriceSetter.cs
@@ -1,5 +1,5 @@
 // Only classes that implement iPriceSetter can set the price of a good
 public interface iPriceSetter
 {
-    void SetPrice(Good good,float newPrice);
+    void SetPrice(Good good,decimal newPrice);
 }
