+++ Assets/Tests/EditMode/CompanyTests.cs
@@ -30,7 +30,7 @@ public class CompanyTests
         var company = ScriptableObject.CreateInstance<Company>();
         company.Initialize("Test Company", CompanyLevelEnum.Beginner);
         var good = Good.CreateInstance("Lemon", new Price_band(1, 3), Rarity_enum.Common);
-        const float trade_price = 3f;
+        const decimal trade_price = 3.0m;
         var expected_cash = 10000 - 3;
         //Act
         company.BuyGood(good, 1, trade_price);
@@ -47,7 +47,7 @@ public class CompanyTests
         var company = ScriptableObject.CreateInstance<Company>();
         company.Initialize("Test Company", CompanyLevelEnum.Beginner);
         var good = Good.CreateInstance("Lemon", new Price_band(1, 3), Rarity_enum.Common);
-        const float trade_price = 3f;
+        const decimal trade_price = 3.0m;
         var expected_inventory_entry = new InventoryEntry(good, 1, trade_price);
         
         //Act
@@ -69,7 +69,7 @@ public class CompanyTests
         var company = ScriptableObject.CreateInstance<Company>();
         company.Initialize("Test Company", CompanyLevelEnum.Beginner);
         var good = Good.CreateInstance("Lemon", new Price_band(1, 3), Rarity_enum.Common);
-        const float trade_price = 3f;
+        const decimal trade_price = 3.0m;
         const int trade_quantity = 10000;
         //Act
         //Assert
@@ -84,7 +84,7 @@ public class CompanyTests
         var good = Good.CreateInstance("Lemon", new Price_band(1, 3), Rarity_enum.Common);
         company.BuyGood(good, 1,good.GetPrice());
         var initial_cash = company.Get_cash();
-        var good_price = 3f;
+        var good_price = 3.0m;
         var expected_cash = initial_cash + good_price;
         
         //Act
@@ -102,7 +102,7 @@ public class CompanyTests
         company.Initialize("Test Company", CompanyLevelEnum.Beginner);
         var good = Good.CreateInstance("Lemon", new Price_band(1, 3), Rarity_enum.Common);
         company.BuyGood(good, 1,good.GetPrice());
-        var good_price = 3f;
+        var good_price = 3.0m;
         //Act
         //Assert
         Assert.Throws<Company_InventoryException>(() => company.SellGood(good, 2, good_price));
